fix: observable leak.
issue #65 explains how to reproduce this and @marukami explains the
reason for this happening beautifully.

given that we want to allow multiple clicks of the button, i think the
right away to address this is just by using `CompositeSubscription`s.
In general, for android this also seems like the way to go.

#67 is also a good potential fix, but i prefer going the CS route.
Seems more "semantic".